% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppOpenMP-package.R
\docType{package}
\name{RcppOpenMP-package}
\alias{RcppOpenMP}
\alias{RcppOpenMP-package}
\title{RcppOpenMP: Using OpenMP with Rcpp}
\description{
Provides an example of parallelizing a C++ routine with 'OpenMP' directly inside an 'Rcpp' function.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/coatless-rd-rcpp/rcpp-openmp}
  \item \url{https://rd-rcpp.thecoatlessprofessor.com/rcpp-openmp/}
  \item Report bugs at \url{https://github.com/coatless-rd-rcpp/rcpp-openmp/issues}
}

}
\author{
\strong{Maintainer}: James Balamuta \email{balamut2@illinois.edu}

Authors:
\itemize{
  \item James Balamuta \email{balamut2@illinois.edu}
}

}
\keyword{internal}
